feature: tratamento de erros

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -56,23 +57,36 @@ public class TopicosController {
 	}
 	
 	@GetMapping("/{id}")
-	public DetalhedoTopicoDto detalhar(@PathVariable Long id) {
-		Topico topico = repository.getOne(id);
-		return new DetalhedoTopicoDto(topico);
+	public ResponseEntity<DetalhedoTopicoDto> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = repository.findById(id);
+		if(topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhedoTopicoDto(topico.get()));
+		}
+		return ResponseEntity.notFound().build();
 	}
 	
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form, UriComponentsBuilder uriComponentsBuilder ){
-		Topico topico = form.atualizar(id, repository);
-		return ResponseEntity.ok(new TopicoDto(topico));
+		Optional<Topico> optional = repository.findById(id);
+		if(optional.isPresent()) {
+			Topico topico = form.atualizar(id, repository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
+		return ResponseEntity.notFound().build();
+
 	}
 	
-	@DeleteMapping ("/{id}")
+	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> atualizar(@PathVariable Long id){
-		repository.deleteById(id);	
-		return ResponseEntity.ok().build();
+		Optional<Topico> optional = repository.findById(id);
+		if(optional.isPresent()) {
+			repository.deleteById(id);	
+			return ResponseEntity.ok().build();
+		}
+		return ResponseEntity.notFound().build();
+
 	}
